Fix offsets and lengths for different blocks. This helps when changing topology.
Also, I think the "process_normals" hack worked by storing existing normals and saving them in the output file. The problem with this approach is that changes in topology would make the model invalid. I tried with one test model and it seems to be working fine. Let's hope this works.